requirements.txt
# Core LiveKit dependencies
livekit-agents[anthropic,silero,turn-detector]>=1.3.0,<1.4.0
livekit-plugins-noise-cancellation>=0.2.0
python-dotenv>=1.0.0
onnxruntime>=1.18.0,<1.19.0
numpy>=1.26.0,<2.0.0
mem0ai>=0.1.0

# Gmail integration dependencies
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.150.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Web search dependencies
httpx>=0.27.0

